nanogenesis/requirements.txt
streamlit>=1.24.0
discord.py>=2.3.0
pydantic>=2.0.0
python-dotenv>=1.0.0
litellm>=1.0.0
aiohttp>=3.8.0
requests>=2.31.0
beautifulsoup4>=4.12.0
duckduckgo-search>=3.8.0
pyyaml>=6.0
Pillow>=10.0.0
pytesseract>=0.3.10
